Refactor back
