+++ src/main/kotlin/nomic/api/RulesAmendments.kt
@@ -1,7 +1,7 @@
 package nomic.api
 
+import nomic.api.models.RulesAmendmentsApiModel
 import nomic.data.EntityNotFoundException
-import nomic.domain.entities.RulesAmendmentsModel
 import nomic.domain.rulesamendments.RuleAmendmentDomain
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
@@ -18,7 +18,7 @@ class RulesAmendments(val ruleAmendmentDomain: RuleAmendmentDomain) {
     // Path to endpoint is api/rules_amendments/ExistingGameId
     @GetMapping("{gameid}", produces = ["application/json;charset=UTF-8"])
     fun getRulesAmendments(@PathVariable(value = "gameid") gameId: String): ResponseEntity<Any> {
-        val rulesAmendments: List<RulesAmendmentsModel> = ruleAmendmentDomain.getRulesAmendments(gameId)
+        val rulesAmendments: List<RulesAmendmentsApiModel> = ruleAmendmentDomain.getRulesAmendments(gameId)
 
         // Return the response object
         return ResponseEntity(rulesAmendments, HttpStatus.OK)
