+++ src/main/kotlin/nomic/domain/entities/RulesAmendmentsModel.kt
@@ -1,36 +1,31 @@
 package nomic.domain.entities
 
 /**
- * The rules and amendments object to be passed back to the caller of the endpoint
+ * Raw representation of data from the repository layer
  *
  * @property ruleId the unique if for the rule
- * @property index the position of the rule within the context of a game
- * @property title the name of this rule
- * @property description the clarifying text for the rule
- * @property mutable flag for mutable/immutable rules
- * @property amendments the list of amendments that are attached to a rule
- * @see[nomic.domain.entities.AmendmentModel]
+ * @property ruleIndex the position of the rule within the context of a game
+ * @property ruleTitle the name of this rule
+ * @property ruleDescription the clarifying text for the rule
+ * @property ruleMutable flag for mutable/immutable rules
+ * @property ruleActive flag to determine rule visibility to the user
+ * @property amendId the primary key for an amendment
+ * @property amendIndex the position of an amendment under a rule
+ * @property amendDescription the context of an amendment
+ * @property amendTitle what an amendment is called
+ * @property amendActive flag for whether an amendment is still in effect
+ * @see[nomic.domain.entities.RulesAmendmentsModel]
  */
 data class RulesAmendmentsModel(
     val ruleId: Int,
-    val index: Int,
-    val title: String,
-    val description: String? = null,
-    val mutable: Boolean,
-    var amendments: MutableList<AmendmentModel>? = mutableListOf()
-)
-
-/**
- * The amendments object to be passed back to the caller of the endpoint
- *
- * @property amendId the primary key for an amendment
- * @property index the position of an amendment under a rule
- * @property description the context of an amendment
- * @property title what an amendment is called
- */
-data class AmendmentModel(
-    val amendId: Int,
-    val index: Int,
-    val description: String,
-    val title: String
+    val ruleIndex: Int,
+    val ruleTitle: String,
+    val ruleDescription: String?,
+    val ruleMutable: Boolean,
+    val ruleActive: Boolean,
+    val amendId: Int?,
+    val amendIndex: Int?,
+    val amendTitle: String?,
+    val amendDescription: String?,
+    val amendActive: Boolean?
 )
