+++ src/main/kotlin/nomic/domain/rulesamendments/IRuleAmendmentDomain.kt
@@ -1,6 +1,6 @@
 package nomic.domain.rulesamendments
 
-import nomic.domain.entities.RulesAmendmentsModel
+import nomic.api.models.RulesAmendmentsApiModel
 
 /**
  * This manages the call and format of the rules and amendments data
@@ -12,5 +12,5 @@ interface IRuleAmendmentDomain {
      * @param gameId The id of the game to collect rule and amendment data on
      * @return The list of [RulesAmendmentsModel][nomic.domain.entities.RulesAmendmentsModel] objects to pass back to the API layer
      */
-    fun getRulesAmendments(gameId: String): MutableList<RulesAmendmentsModel>
+    fun getRulesAmendments(gameId: String): MutableList<RulesAmendmentsApiModel>
 }
