+++ src/main/kotlin/nomic/domain/rulesamendments/RuleAmendmentDomain.kt
@@ -1,8 +1,8 @@
 package nomic.domain.rulesamendments
 
-import nomic.data.dtos.RulesAmendmentsDTO
+import nomic.api.models.AmendmentModel
+import nomic.api.models.RulesAmendmentsApiModel
 import nomic.data.repositories.rulesamendments.RuleAmendmentRepository
-import nomic.domain.entities.AmendmentModel
 import nomic.domain.entities.RulesAmendmentsModel
 import org.springframework.stereotype.Service
 
@@ -19,17 +19,17 @@ import org.springframework.stereotype.Service
 class RuleAmendmentDomain(
     private val ruleAmendmentRepository: RuleAmendmentRepository
 ) : IRuleAmendmentDomain {
-    override fun getRulesAmendments(gameId: String): MutableList<RulesAmendmentsModel> {
+    override fun getRulesAmendments(gameId: String): MutableList<RulesAmendmentsApiModel> {
         val gameIdInt: Int = gameId.toIntOrNull() ?: throw IllegalArgumentException("Please enter a valid GameId!")
 
-        val rulesRaw: List<RulesAmendmentsDTO> = ruleAmendmentRepository.getRulesAmendments(gameIdInt)
-        val rules: MutableList<RulesAmendmentsModel> = mutableListOf()
+        val rulesRaw: List<RulesAmendmentsModel> = ruleAmendmentRepository.getRulesAmendments(gameIdInt)
+        val rules: MutableList<RulesAmendmentsApiModel> = mutableListOf()
         var currId: Int = -1
 
         rulesRaw.forEach { row ->
             if (currId != row.ruleId && row.ruleActive) {
                 currId = row.ruleId
-                rules += RulesAmendmentsModel(
+                rules += RulesAmendmentsApiModel(
                     row.ruleId,
                     row.ruleIndex,
                     row.ruleTitle,
